requirements.txt
aiohttp==3.4.4
aiohttp-cors==0.7.0
aioredis==1.2.0
bcrypt==3.1.4
envparse==0.2.0
trafaret==1.2.0
pyjwt==1.7.0
motor==2.0.0